more selfexplaining upgrade process
At least for me it was not very intuitive what I have to do here. Maybe because I overread "not" and also because I simply likes to upgrade all packages. That was not explained ether, so I didn't consider to just press enter. This hopefully makes it much more understandable.