Patch to honor combinedupgrade properly
In a situation where an AUR package is upgraded, and requires repository make dependencies, the make dependencies would appear in the "Packages to exclude" list. This clutters up the list unnecessarily. The same packages are also listed one step later per AUR package to upgrade.

This removes the repo makedeps from the "Packages to exclude" list when nocombinedupgrade is used (or configured).