Fix incompatible Pacman flags when listing packages
Fixes #2486.

-u and --upgrade cannot be passed into Pacman when -l or --list are
used. Therefore, remove the upgrade arguments before executing the
Pacman flag.

Additionally, the refresh flags are removed because it is unnecessary
and unexpected for the databases to be refreshed without explicitly
passing a refresh flag. A refresh can still be forced by the user by
supplying a refresh flag to yay's command line.